+++ tools/gifconv/gif2hdf.c
@@ -1,3 +1,17 @@
+/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ * Copyright by the Board of Trustees of the University of Illinois.         *
+ * All rights reserved.                                                      *
+ *                                                                           *
+ * This file is part of HDF5.  The full HDF5 copyright notice, including     *
+ * terms governing use, modification, and redistribution, is contained in    *
+ * the files COPYING and Copyright.html.  COPYING can be found at the root   *
+ * of the source code distribution tree; Copyright.html can be found at the  *
+ * root level of an installed copy of the electronic HDF5 document set and   *
+ * is linked from the top-level documents page.  It can also be found at     *
+ * http://hdf.ncsa.uiuc.edu/HDF5/doc/Copyright.html.  If you do not have     *
+ * access to either file, you may request a copy from hdfhelp@ncsa.uiuc.edu. *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
+
 /* #include "hdf.h" */
 #include "gif.h"
 #include <string.h>
@@ -14,7 +28,7 @@ main(int argv , char *argc[])
 	
 	/* replacing int32 with long */
 	long i,ImageCount;
-	long filesize;
+	size_t filesize;
 
 	BYTE *MemGif;
 	BYTE *StartPos;
@@ -55,7 +69,7 @@ main(int argv , char *argc[])
 
 	/* Get the whole file into memory. Mem's much faster than I/O */
 	fseek(fpGif, 0L , 2);
-	filesize = ftell(fpGif);
+	filesize = (size_t)ftell(fpGif);
 	fseek(fpGif, 0L , 0);
 	if (filesize == 0) printf("File Size Zero");
 	if (!(MemGif = StartPos = (BYTE *)malloc(filesize))) {
